requirements.txt
requests
beautifulsoup4
python-telegram-bot
humanize
